@@ -18,40 +18,42 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "^6.0.2",
-    "@angular/cdk": "^6.3.0",
-    "@angular/common": "^6.0.2",
-    "@angular/compiler": "^6.0.2",
-    "@angular/core": "^6.0.2",
-    "@angular/flex-layout": "^6.0.0-beta.16",
-    "@angular/forms": "^6.0.2",
-    "@angular/http": "^6.0.2",
-    "@angular/material": "^6.3.0",
-    "@angular/platform-browser": "^6.0.2",
-    "@angular/platform-browser-dynamic": "^6.0.2",
-    "@angular/router": "^6.0.2",
-    "@ngrx/store": "^6.0.1",
+    "@angular/animations": "^7.0.4",
+    "@angular/cdk": "^7.0.4",
+    "@angular/common": "^7.0.4",
+    "@angular/compiler": "^7.0.4",
+    "@angular/core": "^7.0.4",
+    "@angular/flex-layout": "^7.0.0-beta.19",
+    "@angular/forms": "^7.0.4",
+    "@angular/http": "^7.0.4",
+    "@angular/material": "^7.0.4",
+    "@angular/platform-browser": "^7.0.4",
+    "@angular/platform-browser-dynamic": "^7.0.4",
+    "@angular/pwa": "^0.10.6",
+    "@angular/router": "^7.0.4",
+    "@angular/service-worker": "^7.0.4",
+    "@ngrx/store": "^6.1.2",
     "adal-angular4": "^3.0.1",
     "angularfire2": "^5.0.0-rc.9",
     "core-js": "^2.5.7",
     "firebase": "^5.0.2",
     "hammerjs": "^2.0.8",
-    "rxjs": "^6.1.0",
+    "rxjs": "^6.3.3",
     "rxjs-compat": "^6.2.1",
     "web-animations-js": "^2.3.1",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.6.3",
-    "@angular/cli": "^6.0.3",
-    "@angular/compiler-cli": "^6.0.2",
-    "@angular/language-service": "^6.0.2",
+    "@angular/cli": "^7.0.6",
+    "@angular/compiler-cli": "^7.0.4",
+    "@angular/language-service": "^7.0.4",
     "@types/jasmine": "~2.8.7",
     "@types/jasminewd2": "~2.0.2",
     "@types/karma": "^1.7.3",
     "@types/node": "^10.1.0",
     "codelyzer": "^4.3.0",
-    "firebase-tools": "^3.18.6",
+    "firebase-tools": "^6.1.0",
     "husky": "^0.14.3",
     "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
@@ -67,7 +69,7 @@
     "standard-version": "^4.3.0",
     "ts-node": "~6.1.1",
     "tslint": "^5.10.0",
-    "typescript": ">=2.7.0 <2.8.0",
+    "typescript": "3.1.6",
     "yarn-run-all": "^3.1.1"
   },
   "repository": {
