@@ -4,6 +4,6 @@ export const config = {
     POPULAR_MOVIES_URL: 'https://api.themoviedb.org/3/movie/popular?',
     CASTCREW_URL: 'https://api.themoviedb.org/3/movie',
     API_KEY: 'api_key=397b8dec59143fda45d64baf789161d6',
-    MOVIE_IMAGE_URL: 'http://image.tmdb.org/t/p/w500/'
+    MOVIE_IMAGE_URL: 'https://image.tmdb.org/t/p/w500/'
   }
 };
